Fixed issue from sonarqube related to Optional().get()

+++ src/main/java/ch/uzh/ifi/hase/soprafs24/service/RoomService.java
@@ -72,7 +72,9 @@ public class RoomService {
             newRoom.setRoomPlayersList(newRoom.getRoomPlayersList());
 
             newRoom = roomRepository.save(newRoom);
-
+            if(userRepository.findById(newRoom.getRoomOwnerId()).isEmpty()){
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Room owner not found");
+            }
             User roomOwner = userRepository.findById(newRoom.getRoomOwnerId()).get();
             roomOwner.setPlayerStatus(PlayerStatus.READY);
             roomOwner.setInRoomId(newRoom.getRoomId());
@@ -138,6 +140,9 @@ public class RoomService {
             if (gameRepository.findByRoomId(room.getRoomId()).isPresent()){
                 System.out.println(room.getRoomId());
                 System.out.println(gameRepository.findByRoomId(room.getRoomId()).isPresent());
+                if(gameRepository.findByRoomId(room.getRoomId()).isEmpty()){
+                    throw new RuntimeException("Game not found");
+                }
                 gameRepository.delete(gameRepository.findByRoomId(room.getRoomId()).get());
             }
             roomRepository.delete(room);
@@ -145,6 +150,9 @@ public class RoomService {
         else{
             if(room.getRoomOwnerId().equals(user.getId()) && room.getRoomPlayersList().size() > 1) {
                 room.setRoomOwnerId(room.getRoomPlayersList().get(1));
+                if(userRepository.findById(room.getRoomPlayersList().get(1)).isEmpty()){
+                    throw new RuntimeException("New room owner not found");
+                }
                 User newOwner = userRepository.findById(room.getRoomPlayersList().get(1)).get();
                 newOwner.setPlayerStatus(PlayerStatus.READY);
                 userRepository.save(newOwner);
